Split integral_BdotN into a separate source file
