+++ src/simsoptpp/integral_BdotN.h
@@ -1,0 +1,4 @@
+#include "xtensor-python/pyarray.hpp"
+typedef xt::pyarray<double> PyArray;
+
+double integral_BdotN(PyArray& Bcoil, PyArray& Btarget, PyArray& n, std::string definition);
